requirements.txt
# FastAPI
fastapi
uvicorn
python-multipart
aiofiles

# Audio y vídeo
ffmpeg-python
moviepy
pydub
librosa
soundfile
scipy
numpy

# IA y procesamiento
openai
torch
transformers
torchaudio

# Clonación de voz externa
elevenlabs

# Config
python-dotenv

# Base de datos y autenticación
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]

# Cliente HTTP
httpx

# Sistema de pagos
stripe